Extract common tests to base class
