+++ src/test/php/web/session/unittest/InFileSystemTest.class.php
@@ -4,7 +4,6 @@ use io\Folder;
 use lang\Environment;
 use lang\IllegalArgumentException;
 use web\session\InFileSystem;
-use web\session\ISession;
 
 class InFileSystemTest extends SessionsTest {
   private static $dir;
@@ -20,14 +19,12 @@ class InFileSystemTest extends SessionsTest {
     self::$dir->exists() && self::$dir->unlink();
   }
 
-  #[@test]
-  public function can_create() {
-    new InFileSystem();
-  }
+  /** @return web.session.Sessions */
+  protected function fixture() { return new InFileSystem(self::$dir); }
 
   #[@test]
-  public function can_create_with_dir() {
-    new InFileSystem(self::$dir);
+  public function can_create_without_argument() {
+    new InFileSystem();
   }
 
   #[@test, @expect(IllegalArgumentException::class)]
@@ -36,64 +33,9 @@ class InFileSystemTest extends SessionsTest {
   }
 
   #[@test]
-  public function create() {
-    $sessions= new InFileSystem(self::$dir);
-    $this->assertInstanceOf(ISession::class, $sessions->create());
-  }
-
-  #[@test]
   public function session_identifiers_consist_of_32_lowercase_hex_digits() {
-    $sessions= new InFileSystem(self::$dir);
+    $sessions= $this->fixture();
     $id= $sessions->create()->id();
     $this->assertTrue((bool)preg_match('/^[a-f0-9]{32}$/i', $id), $id);
-  }
-
-  #[@test]
-  public function read_write() {
-    $session= (new InFileSystem(self::$dir))->create();
-    $session->register('name', 'value');
-    $this->assertEquals('value', $session->value('name'));
-  }
-
-  #[@test]
-  public function read_non_existant() {
-    $session= (new InFileSystem(self::$dir))->create();
-    $this->assertNull($session->value('name'));
-  }
-
-  #[@test]
-  public function read_non_existant_returns_default() {
-    $session= (new InFileSystem(self::$dir))->create();
-    $this->assertEquals('Default value', $session->value('name', 'Default value'));
-  }
-
-  #[@test]
-  public function remove() {
-    $session= (new InFileSystem(self::$dir))->create();
-    $session->register('name', 'value');
-    $session->remove('name');
-    $this->assertNull($session->value('name'));
-  }
-
-  #[@test]
-  public function read_write_with_two_session_instances() {
-    $sessions= new InFileSystem(self::$dir);
-
-    $session= $sessions->create();
-    $session->register('name', 'value');
-    $session->transmit($this->response());
-
-    $session= $sessions->open($this->request('session='.$session->id()));
-    $value= $session->value('name');
-
-    $this->assertEquals('value', $value);
-  }
-
-  #[@test]
-  public function destroy() {
-    $session= (new InFileSystem(self::$dir))->create();
-    $session->register('name', 'value');
-    $session->destroy();
-    $this->assertFalse($session->valid());
   }
 }
